fix(customcheckboxinput): fixed formik handling of Chakra's checkbox
This was a bug which happened when the checkbox was clicked. It would give a type error. Formik was
applying an onChange function in the CheckBoxgroup component, which was triggered two times when the
user clicked a checkbox, giving the type error. To solve the problem I just added only the values to
the CheckboxGroup and then all the Field properties to the Checkbox